+++ src-gen/com/holycityaudio/SpinCAD/CADBlocks/PatternGeneratorCADBlock.java
@@ -10,7 +10,7 @@
  * 
  *   This program is distributed in the hope that it will be useful, 
  *   but WITHOUT ANY WARRANTY; without even the implied warranty of 
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the 
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
  *   GNU General Public License for more details. 
  * 
  *   You should have received a copy of the GNU General Public License 
@@ -81,7 +81,7 @@
 			hold = sfxb.allocateReg();
 			if(this.getPin("Rate").isConnected() == true) {
 			sfxb.skip(RUN, 1);
-			sfxb.loadRampLFO(0, 20, 4096);
+			sfxb.loadRampLFO((int) 0, (int) 20, (int) 4096);
 			sfxb.loadAccumulator(rate);
 			sfxb.mulx(rate);
 			sfxb.mulx(rate);
