+++ src-gen/com/holycityaudio/SpinCAD/CADBlocks/ga_demo_vibratoCADBlock.java
@@ -10,7 +10,7 @@
  * 
  *   This program is distributed in the hope that it will be useful, 
  *   but WITHOUT ANY WARRANTY; without even the implied warranty of 
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the 
+ *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
  *   GNU General Public License for more details. 
  * 
  *   You should have received a copy of the GNU General Public License 
@@ -147,8 +147,8 @@
 			sfxb.skip(RUN, 4);
 			sfxb.writeRegister(lp1, 0);
 			sfxb.writeRegister(lp2, 0);
-			sfxb.loadSinLFO(SIN0, 12, 100);
-			sfxb.loadSinLFO(SIN1, 0, 0);
+			sfxb.loadSinLFO((int) SIN0,(int) 12, (int) 100);
+			sfxb.loadSinLFO((int) SIN1,(int) 0, (int) 0);
 			sfxb.readRegister(adcl, 0.5);
 			sfxb.readRegister(adcr, 0.5);
 			sfxb.FXwriteDelay("vdel", 0, 1);
